Move ViewModel interface into the shared components
