+++ src/shared-components/ViewModel.ts
